requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-jsonapi==7.1.0
idna==3.7
inflection==0.5.1
mysql==0.0.3
mysqlclient==2.2.4
pillow==10.4.0
razorpay==1.4.2
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
